+++ prediction_market_agent_tooling/markets/omen/omen_contracts.py
@@ -893,30 +893,6 @@ class OmenThumbnailMapping(ContractOnGnosisChain):
         )
 
 
-class CowGPv2SettlementContract(ContractOnGnosisChain):
-    # Contract ABI taken from https://github.com/cowprotocol/cow-sdk/blob/main/abi/GPv2Settlement.json.
-    abi: ABI = abi_field_validator(
-        os.path.join(
-            os.path.dirname(os.path.realpath(__file__)),
-            "../../abis/gvp2_settlement.abi.json",
-        )
-    )
-
-    def setPreSignature(
-        self,
-        api_keys: APIKeys,
-        orderId: HexBytes,
-        signed: bool,
-        web3: Web3 | None = None,
-    ) -> None:
-        self.send(
-            api_keys=api_keys,
-            function_name="setPreSignature",
-            function_params=[orderId, signed],
-            web3=web3,
-        )
-
-
 class CollateralTokenChoice(str, Enum):
     wxdai = "wxdai"
     sdai = "sdai"
